Add `ITransformable` interface to support transformation actions as traits.
* Make `Url` ITransformable`
* Make LayerContainer.Builder `ITransformable`
* Update layer tests to use new  syntax.
